dashboard/requirements.txt
streamlit
kafka-python
pandas
altair
watchdog